requirements.txt
Flask==2.2.2
Werkzeug==2.3.8
gunicorn==20.1.0
requests
Flask-Cors
psycopg2-binary
redis>=4.0.0
rq>=1.15.0
openai>=1.0.0
anthropic>=0.18.0
b2sdk>=1.24.0
boto3>=1.34.0
python-dotenv==0.21.0
Pillow>=10.0.0
PyJWT>=2.0.0
